requirements.txt
SQLAlchemy~=1.4.45
alembic~=1.9.1
aiogram~=2.24
python-dotenv~=0.21.0
loguru~=0.6.0
gino~=1.0.1